+++ storybook/lib/stories/nested_overlays.dart
@@ -3,33 +3,33 @@ import 'package:flutter/widgets.dart';
 import 'package:optimus/optimus.dart';
 import 'package:storybook/stories/nonmodal_wrapper.dart';
 import 'package:storybook/stories/search_field.dart';
-import 'package:storybook/stories/select.dart';
+import 'package:storybook/stories/select_input.dart';
 import 'package:storybook_flutter/storybook_flutter.dart';
 
 final Story nestedSelectStory = Story(
   section: 'Nested overlays',
   name: 'Nested Select',
-  builder: (_, k) => _Content(SelectStory(k)),
+  builder: (_, k) => _Content((_) => SelectInputStory(k)),
 );
 
 final Story nestedSearchStory = Story(
   section: 'Nested overlays',
   name: 'Nested Search',
-  builder: (_, k) => _Content(SearchStory(k)),
+  builder: (_, k) => _Content((_) => SearchStory(k)),
 );
 
 final Story nestedNonModalDialogStory = Story(
   section: 'Nested overlays',
   name: 'Nested Non-modal dialog',
   builder: (_, k) => _Content(
-    Center(child: NonModalWrapper(child: NonModalDialogStory(k))),
+    (_) => Center(child: NonModalWrapper(child: NonModalDialogStory(k))),
   ),
 );
 
 class _Content extends StatelessWidget {
-  const _Content(this.child, {Key? key}) : super(key: key);
+  const _Content(this.contentBuilder, {Key? key}) : super(key: key);
 
-  final Widget child;
+  final WidgetBuilder contentBuilder;
 
   @override
   Widget build(BuildContext context) => Column(
@@ -45,7 +45,7 @@ class _Content extends StatelessWidget {
                   case 'initialRoute':
                     builder = (context) => ConstrainedBox(
                           constraints: const BoxConstraints(maxWidth: 400),
-                          child: child,
+                          child: contentBuilder(context),
                         );
                     break;
                 }