+++ storybook/lib/stories/select_input.dart
@@ -3,35 +3,45 @@ import 'package:optimus/optimus.dart';
 import 'package:storybook/utils.dart';
 import 'package:storybook_flutter/storybook_flutter.dart';
 
-final Story selectStory = Story(
-  name: 'Select',
+final Story selectInputStory = Story(
+  name: 'Select Input',
   builder: (_, k) => ConstrainedBox(
     constraints: const BoxConstraints(maxWidth: 400),
-    child: SelectStory(k),
+    child: SelectInputStory(k),
   ),
 );
 
-class SelectStory extends StatefulWidget {
-  const SelectStory(this.knobs, {Key? key}) : super(key: key);
+class SelectInputStory extends StatefulWidget {
+  const SelectInputStory(this.knobs, {Key? key}) : super(key: key);
 
   final KnobsBuilder knobs;
 
   @override
-  _SelectStoryState createState() => _SelectStoryState();
+  _SelectInputStoryState createState() => _SelectInputStoryState();
 }
 
-class _SelectStoryState extends State<SelectStory> {
-  int? _selectedValue;
+class _SelectInputStoryState extends State<SelectInputStory> {
+  String? _selectedValue;
+  String _searchToken = '';
+
+  void _onTextChanged(String text) {
+    setState(() {
+      _searchToken = text.toLowerCase();
+    });
+  }
 
   @override
   Widget build(BuildContext context) {
     final k = widget.knobs;
-    return OptimusSelect<int>(
+
+    return OptimusSelectInput<String>(
       value: _selectedValue,
       isEnabled: k.boolean(label: 'Enabled', initial: true),
       isRequired: k.boolean(label: 'Required'),
       prefix:
           k.boolean(label: 'Prefix') ? const Icon(OptimusIcons.search) : null,
+      onTextChanged: k.boolean(label: 'Searchable') ? _onTextChanged : null,
+      onChanged: (i) => setState(() => _selectedValue = i),
       size: k.options(
         label: 'Size',
         initial: OptimusWidgetSize.large,
@@ -42,15 +52,23 @@ class _SelectStoryState extends State<SelectStory> {
       caption: Text(k.text(label: 'Caption', initial: '')),
       secondaryCaption: Text(k.text(label: 'Secondary caption', initial: '')),
       error: k.text(label: 'Error', initial: ''),
-      items: Iterable<int>.generate(10)
-          .map((i) => ListDropdownTile<int>(
-                value: i,
-                title: Text('Dropdown tile #$i'),
-                subtitle: Text('Subtitle #$i'),
+      items: _characters
+          .where((e) => e.toLowerCase().contains(_searchToken))
+          .map((e) => ListDropdownTile<String>(
+                value: e,
+                title: Text(e),
+                subtitle: Text(e.toUpperCase()),
               ))
           .toList(),
-      builder: (context, option) => Text('Dropdown tile #$option'),
-      onItemSelected: (i) => setState(() => _selectedValue = i),
+      builder: (option) => option,
     );
   }
 }
+
+const _characters = [
+  'Jon Snow',
+  'Ned Stark',
+  'Robb Stark',
+  'Sansa Stark',
+  'Daenerys Targaryen',
+];